// frontend/src/app/pages/model/PlanActPcd.ts
export class PlanActPcd {
  id: number;
  year: number;
  week: number;
  ecartEuro: number;
  besoin: string;
  cause: string;
  action: string;
  qui: string;
  dateDebut: Date;
  dateObjectif: Date;
  dateRealisation: Date;
  actionEfficace: string;
  generalisationPosssible: string;
  comment: string;
  atelier: string;

  constructor() {}
}